requirements.txt
pandas
numpy
mac_vendor_lookup
requests.html
googlesearch-python
